CPP02/ex03 in progress

+++ CPP02/ex03/bsp.cpp
@@ -6,7 +6,7 @@
 /*   By: ddavlety <ddavlety@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/15 11:11:58 by ddavlety          #+#    #+#             */
-/*   Updated: 2024/03/15 12:32:42 by ddavlety         ###   ########.fr       */
+/*   Updated: 2024/05/05 17:36:36 by ddavlety         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 
 float	triangle_area(Point const a, Point const b, Point const c)
 {
-	return std::abs(((((a.get_x() * (b.get_y() - c.get_y()))
-			+ (b.get_x() * (c.get_y() - a.get_y()))
+	return std::abs(((((a.get_x().toFloat() * (b.get_y().toFloat() - c.get_y().toFloat()))
+			+ (b.get_x().toFloat() * (c.get_y().toFloat() - a.get_y().toFloat()))
 			+ (c.get_x() * (a.get_y() - b.get_y()))) / 2)));
 }
